refactor: 로또 담첨 결과 확인 메서드를 Rank로 이동
- 로또 당첨 결과 확인 Lotto에서 Rank로 이동
